backend/requirements.txt
djangorestframework
djoser
Django
django-filter
django-extensions
reportlab
django-import-export
psycopg2
